requirements.txt
Django~=4.2.11
numpy~=1.26.4
tensorflow~=2.15.0
djangorestframework~=3.15.1
keras~=3.1.1
pillow~=10.3.0
opencv-python~=4.9.0.80
scipy~=1.13.0
scikit-image~=0.22.0
requests~=2.31.0